LoggerParser uses a special function to get the log file which may set different paths (for example to a working directory). Thus just checking that fileName exists may not work. So moved the check for existing file into XMLParser.getFilePrintWriter(). This has the added benefit that all loggers use this function and thus all have the overwriting behaviour.
Also added a check box to the BEAST options dialog box.

+++ trunk/src/dr/xml/XMLParser.java
@@ -28,6 +28,7 @@ package dr.xml;
 import dr.inference.model.Likelihood;
 import dr.inference.model.Model;
 import dr.inference.model.Parameter;
+import dr.inferencexml.loggers.LoggerParser;
 import dr.util.FileHelpers;
 import dr.util.Identifiable;
 import org.w3c.dom.*;
@@ -349,11 +350,29 @@ public class XMLParser {
      * @throws XMLParseException if file can't be created for some reason
      */
     public static PrintWriter getFilePrintWriter(XMLObject xo, String parserName) throws XMLParseException {
+
         if (xo.hasAttribute(FileHelpers.FILE_NAME)) {
 
             final String fileName = xo.getStringAttribute(FileHelpers.FILE_NAME);
-
             final File logFile = FileHelpers.getFile(fileName);
+
+            boolean allowOverwrite = false;
+
+            if (xo.hasAttribute(LoggerParser.ALLOW_OVERWRITE_LOG)) {
+                allowOverwrite = xo.getBooleanAttribute(LoggerParser.ALLOW_OVERWRITE_LOG);
+            }
+
+            // override with a runtime set System Property
+            if (System.getProperty("allow.overwrite") != null) {
+                allowOverwrite = Boolean.parseBoolean(System.getProperty("allow.overwrite", "false"));
+            }
+
+            if (logFile.exists() && !allowOverwrite) {
+                throw new XMLParseException("\nThe log file " + fileName + " already exists in the working directory." +
+                        "\nTo allow it to be overwritten, use the '-overwrite' command line option when running" +
+                        "\nBEAST or select the option in the Run Options dialog box as appropriate.");
+            }
+
 
             try {
                 return new PrintWriter(new FileOutputStream(logFile));
